app/requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.116.1
h11==0.14.0
pytest
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pydantic==2.11.7
pydantic_core==2.33.2
pydantic[email]>=2.0.0
requests==2.32.4
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
passlib
psycopg2-binary==2.9.9
python-dotenv
python-jose
